refactor: move validation to lib, rename some files
