+++ src/app/core/constants/app.constant.ts
@@ -1,5 +1,4 @@
 export class ApplicationConstants {
     public static readonly deviceIdHeader = 'x-device-uuid';
-    public static readonly PAYLOAD_VALIDATION_ERRORS = 'payload_validator_error';
     public static readonly authorization = 'authorization';
 }
