+++ src/app/core/helpers/errors.ts
@@ -3,26 +3,9 @@ import { Logger } from '@core/utils';
 import { translate } from '@lib/translation';
 import { NextFunction, Request, Response } from 'express';
 import { NetworkStatus } from './network-status';
-import { ApplicationConstants } from '@core/constants';
-import stage from './stage';
 
 const log = new Logger('Errors');
 
-export enum Errors {
-    CastError = 'CastError',
-    AssertionError = 'AssertionError',
-    ERR_AMBIGUOUS_ARGUMENT = 'ERR_AMBIGUOUS_ARGUMENT',
-    MongoError = 'MongoError',
-    ErrorResponse = 'ErrorResponse',
-    SuccessResponse = 'SuccessResponse',
-    JsonWebTokenError = 'JsonWebTokenError',
-    TokenExpiredError = 'TokenExpiredError',
-    NotBeforeError = 'NotBeforeError',
-    ValidationError = 'ValidationError',
-    MulterError = 'MulterError',
-    PayloadTooLargeError = 'PayloadTooLargeError:',
-    StrictModeError = 'StrictModeError'
-}
 
 export class ErrorHandling {
 
@@ -55,52 +38,6 @@ export class ErrorHandling {
         });
     }
 
-    public static catchError(error: any) {
-        const response = new ErrorResponse(error.message,
-            isNaN(error.code)
-                ? NetworkStatus.INTERNAL_SERVER_ERROR
-                : error.code
-        );
-
-        switch (error.name) {
-            case Errors.AssertionError:
-            case Errors.ERR_AMBIGUOUS_ARGUMENT:
-                response.code = stage.production ? NetworkStatus.BAD_REQUEST : response.code;
-            case Errors.StrictModeError:
-                response.message = translate('over_posting_is_not_allowed');
-                response.code = NetworkStatus.BAD_REQUEST;
-                break;
-            case Errors.CastError:
-                response.message = translate('invalid_syntax', { name: error.path });
-                response.code = NetworkStatus.BAD_REQUEST;
-                break;
-            case Errors.ValidationError:
-                // Mongoose validation error
-                response.code = NetworkStatus.BAD_REQUEST;
-                break;
-            case ApplicationConstants.PAYLOAD_VALIDATION_ERRORS:
-                // class validator validation error
-                response.code = NetworkStatus.BAD_REQUEST;
-                break;
-            case Errors.TokenExpiredError:
-                response.message = translate('jwt_expired');
-                response.code = NetworkStatus.UNAUTHORIZED;
-                break;
-            case Errors.JsonWebTokenError:
-                response.message = translate(stage.production ? 'jwt_expired' : error.message);
-                response.code = NetworkStatus.UNAUTHORIZED;
-                break;
-            case Errors.MulterError:
-                response.code = NetworkStatus.BAD_REQUEST;
-                break;
-            case Errors.PayloadTooLargeError:
-                response.code = NetworkStatus.BAD_REQUEST;
-                response.message = 'request entity too large';
-                break;
-            default:
-        }
-        return response;
-    }
 
     public static notFound(req: Request, res: Response) {
         const error = new ErrorResponse(
