+++ src/app/core/helpers/index.ts
@@ -1,7 +1,6 @@
 export * from './hash';
 export * from './errors';
 export * from './stage';
-export * from './errors';
 export * from './network-status';
 export * from './constants';
 export * from './route';
