+++ src/app/core/response/error.response.ts
@@ -1,4 +1,6 @@
-import { HttpResponse } from './response'; import { NetworkStatus } from '@core/helpers'; import { translate } from '@lib/translation';
+import { NetworkStatus } from '@core/helpers/network-status';
+import { translate } from '@lib/translation';
+import { HttpResponse } from './generic.response';
 
 export class ErrorResponse extends HttpResponse {
     constructor(message: string, code = NetworkStatus.BAD_REQUEST) {
