+++ src/app/core/response/generic.response.ts
@@ -1,6 +1,6 @@
 import { NetworkStatus } from '@core/helpers/network-status';
 
-export abstract class HttpResponse extends Error {
+export class HttpResponse extends Error {
     public status: string;
     constructor(public code: number) {
         super();
