+++ src/app/core/response/index.ts
@@ -1,5 +1,11 @@
-import { ErrorResponse } from './error.response';
+export * from './error.response';
+export * from './result';
+export * from './success.response';
+export * from './generic.response';
+
+
 import { NetworkStatus } from '@core/helpers';
+import { ErrorResponse } from './error.response';
 import { SuccessResponse } from './success.response';
 
 export namespace Responses {
@@ -30,6 +36,3 @@ export namespace Responses {
     }
 
 }
-export * from './result';
-export * from './error.response';
-export * from './success.response';
