+++ src/app/shared/common/validation.ts
@@ -1,8 +1,6 @@
-import { validateOrReject, ValidationError, IsString, IsMongoId, Matches, IsEmail, IsJWT } from 'class-validator';
-import { ApplicationConstants } from '@core/constants';
-import { Type, AppUtils } from '@core/utils';
-import { NextFunction, Response, Request } from 'express';
 import { PrimaryKey } from '@lib/mongoose';
+import { IsEmail, IsJWT, IsMongoId, IsString, Matches } from 'class-validator';
+import { validate, PayloadValidator } from '@lib/validation';
 
 export class ValidationPatterns {
     public static EmailValidation = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
@@ -11,7 +9,7 @@ export class ValidationPatterns {
 }
 
 export function isValidId() {
-    class IdValidator {
+    class IdValidator extends PayloadValidator {
         @IsMongoId({ message: 'id_not_valid' }) id: string = null;
     }
     return validate(IdValidator, 'params');
@@ -39,27 +37,4 @@ export class TokenValidator {
 
 export class PrimaryIDValidator {
     @IsMongoId() id: PrimaryKey = null;
-}
-
-export async function validatePayload<T>(payload: T, message?: string) {
-    try {
-        await validateOrReject(payload);
-    } catch (validationErrors) {
-        // TODO: Add custom validation error
-        const errorConstraints = (validationErrors[0] as ValidationError).constraints;
-        const error = new Error(message ?? Object.values(errorConstraints)[0]);
-        error.name = ApplicationConstants.PAYLOAD_VALIDATION_ERRORS;
-        throw error;
-    }
-}
-
-export abstract class PayloadValidator { }
-
-export function validate<T extends PayloadValidator>(validator: Type<T>, type: 'body' | 'query' | 'params' | 'headers' | 'queryPolluted' = 'body', message?: string) {
-    return async (req: Request, res: Response, next: NextFunction) => {
-        const validatee = new validator();
-        AppUtils.strictAssign(validatee, req[type]);
-        await validatePayload(validatee, message);
-        next();
-    };
 }
