+++ src/lib/mongoose/entity.ts
@@ -1,12 +1,12 @@
 import { BaseSchema } from '@core/database';
-import { Type } from '@core/utils';
 import { locate, Locator } from '@lib/locator';
 import { Document as MongooseDocument, model, Model, SchemaOptions } from 'mongoose';
 import 'reflect-metadata';
 import { generateModelMetadataKey } from '.';
+import { Type } from '@lib/utils';
 
 export function Entity(name?: string, options: SchemaOptions = {}) {
-    return function(constructor) {
+    return function (constructor) {
         const metadataKey = generateModelMetadataKey(constructor);
         const fields = Reflect.getMetadata(metadataKey, constructor);
         Reflect.deleteMetadata(metadataKey, constructor);
