+++ src/lib/utils.ts
@@ -1,0 +1,3 @@
+
+export type Type<T> = new (...args: any[]) => T;
+export type Parameter<T extends (args: any) => any> = T extends (args: infer P) => any ? P : never;
